Final namespace changes.
