+++ src/Demo/NetExDemo/NetEx.Demo.csproj
@@ -47,7 +47,7 @@
 
 	<ItemGroup>
 		<ProjectReference Include="..\..\NetEx.Hooks\NetEx.Hooks.csproj" />
-		<ProjectReference Include="..\..\NetEx.WinForms.Dialogs\NetEx.WinForms.Dialogs.csproj" />
+		<ProjectReference Include="..\..\NetEx.WinForms.Dialogs\NetEx.Dialogs.WinForms.csproj" />
 	</ItemGroup>
 
 	<ItemGroup Condition="$(TargetFramework.ToLower().Contains('net20'))">
