+++ src/NetEx.WinForms.Dialogs/Interop/CREDUI_INFO.cs
@@ -2,7 +2,7 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Runtime.InteropServices;
 
-namespace NetEx.Windows.Forms.Internal
+namespace NetEx.Dialogs.WinForms.Interop
 {
     /// <summary>
     /// The CREDUI_INFO structure is used to pass information to the CredUIPromptForCredentials function that creates a dialog box used to obtain credentials information.
