+++ src/NetEx.WinForms.Dialogs/Interop/WindowsProgressDialog.cs
@@ -1,6 +1,6 @@
 ﻿using System.Runtime.InteropServices;
 
-namespace NetEx.Windows.Forms.Internal
+namespace NetEx.Dialogs.WinForms.Interop
 {
     [ComImport]
     [Guid("F8383852-FCD3-11d1-A6B9-006097DF5BD4")]
