+++ src/NetEx.sln
@@ -7,7 +7,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Demo", "Demo", "{02EA681E-C
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetEx.Demo", "Demo\NetExDemo\NetEx.Demo.csproj", "{91937276-F353-4B0A-BBA8-92FE3AD58E6D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetEx.WinForms.Dialogs", "NetEx.WinForms.Dialogs\NetEx.WinForms.Dialogs.csproj", "{7BDEE9A0-6915-D481-CAE9-3D3247AFCC8A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetEx.Dialogs.WinForms", "NetEx.WinForms.Dialogs\NetEx.Dialogs.WinForms.csproj", "{7BDEE9A0-6915-D481-CAE9-3D3247AFCC8A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetEx.Hooks", "NetEx.Hooks\NetEx.Hooks.csproj", "{DCEDCDF4-3674-C1E2-BAA8-CDF577AAD057}"
 EndProject
